Add capitalize function back for polls. Stop deleting these messages. Check that message exists before deletion for say commands.

+++ commands/community/say.js
@@ -13,7 +13,7 @@ module.exports = {
       )
     }
 
-    message.delete()
+    if (message) message.delete()
     messageChannel.send(args.join(' '))
   },
 }
